+++ entity-view/impl/src/main/java/com/blazebit/persistence/view/impl/objectbuilder/transformer/SingularMultisetTupleTransformerFactory.java
@@ -65,7 +65,7 @@ public class SingularMultisetTupleTransformerFactory implements TupleTransformer
     @Override
     public TupleTransformer create(ParameterHolder<?> parameterHolder, Map<String, Object> optionalParameters, EntityViewConfiguration entityViewConfiguration) {
         if (!entityViewConfiguration.hasSubFetches(attributePath)) {
-            return new NullTupleTransformer(template, startIndex);
+            return NullTupleTransformer.forMultiset(startIndex);
         }
         if (mapping != null) {
             if (parameterHolder instanceof FullQueryBuilder<?, ?>) {