+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Changes that happened in releases
 * Use MySQL date/time functions also on MariaDB
 * Allow adding custom GraphQL deserialization types and auto-add `java.time.Ser` for time types
 * Add missing quotes around order direction rendering of `JPQLNextExpressions.groupConcat()`
+* Ensure dynamic fetching of correlated collections works with multiset fetching
 
 ### Backwards-incompatible changes